+++ Proxytrace.Domain/TestCase/Internal/TestCaseGenerator.cs
@@ -25,5 +25,6 @@ internal class TestCaseGenerator : DomainEntityGenerator<ITestCase>
     public override async Task<ITestCase> GenerateAsync(CancellationToken cancellationToken = default)
         => factory(
             input: await conversationGenerator.CreateAsync(cancellationToken),
-            expectedOutput: await assistantMessageGenerator.CreateAsync(cancellationToken));
+            expectedOutput: await assistantMessageGenerator.CreateAsync(cancellationToken),
+            sourceAgentCallId: null);
 }